hotfix 20.0.2

+++ malloc_free/0-create_array.c
@@ -10,7 +10,7 @@
 
 char *create_array(unsigned int size, char c)
 {
-	int i;
+	unsigned int i;
 	char *array;
 
 	if (size == 0)
@@ -19,7 +19,7 @@ char *create_array(unsigned int size, char c)
 
 	if(array == NULL)
 		return (NULL);
-	for (i = 0; i < (int)size; i++)
+	for (i = 0; i < size; i++)
 		array[i] = c;
 	return (array);
 }
